--- a/compiler/src/lib.rs
+++ b/compiler/src/lib.rs
@@ -1,6 +1,7 @@
 // Keep these in sync with `main.rs`!
 #![feature(async_closure)]
 #![feature(box_patterns)]
+#![feature(entry_insert)]
 #![feature(let_chains)]
 #![feature(never_type)]
 #![feature(try_trait_v2)]