--- a/compiler/src/main.rs
+++ b/compiler/src/main.rs
@@ -1,6 +1,7 @@
 // Keep these in sync with `lib.rs`!
 #![feature(async_closure)]
 #![feature(box_patterns)]
+#![feature(entry_insert)]
 #![feature(let_chains)]
 #![feature(never_type)]
 #![feature(try_trait_v2)]