--- a/compiler/src/vm/builtin_functions.rs
+++ b/compiler/src/vm/builtin_functions.rs
@@ -10,6 +10,7 @@ use itertools::Itertools;
 use num_bigint::BigInt;
 use num_integer::Integer;
 use num_traits::ToPrimitive;
+use paste::paste;
 use std::{ops::Deref, str::FromStr};
 use tracing::{info, span, Level};
 use unicode_segmentation::UnicodeSegmentation;
@@ -130,7 +131,7 @@ macro_rules! unpack {
             let ( $( $arg, )+ ): ( $( UnpackedData<$type>, )+ ) = ( $(
                 UnpackedData {
                     address: $arg,
-                    data: $heap.get($arg).data.clone().try_into()?,
+                    data: (&$heap.get($arg).data).try_into()?,
                 },
             )+ );
 
@@ -151,12 +152,18 @@ macro_rules! unpack_and_later_drop {
             let ( $( $arg, )+ ): ( $( UnpackedData<$type>, )+ ) = ( $(
                 UnpackedData {
                     address: $arg,
-                    data: $heap.get($arg).data.clone().try_into()?,
+                    data: (&$heap.get($arg).data).try_into()?,
                 },
             )+ );
 
+            // Structs are called `struct_`, so we sometimes generate
+            // identifiers containing a double underscore.
+            #[allow(non_snake_case)]
+            $( let paste!([< $arg _address >]) = $arg.address; )+
+
             let result = $body;
-            $( $heap.drop($arg.address); )+
+
+            $( $heap.drop(paste!([< $arg _address >])); )+
             result.into()
         }
     };
@@ -164,7 +171,7 @@ macro_rules! unpack_and_later_drop {
 
 impl Heap {
     fn channel_create(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |capacity: Int| {
+        unpack_and_later_drop!(self, args, |capacity: &Int| {
             match capacity.value.clone().try_into() {
                 Ok(capacity) => CreateChannel { capacity },
                 Err(_) => return Err("You tried to create a channel with a capacity that is either negative or bigger than the maximum usize.".to_string()),
@@ -173,7 +180,7 @@ impl Heap {
     }
 
     fn channel_send(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |port: SendPort, packet: Any| {
+        unpack_and_later_drop!(self, args, |port: &SendPort, packet: Any| {
             let mut heap = Heap::default();
             let address = self.clone_single_to_other_heap(&mut heap, packet.address);
             Send {
@@ -184,7 +191,7 @@ impl Heap {
     }
 
     fn channel_receive(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |port: ReceivePort| {
+        unpack_and_later_drop!(self, args, |port: &ReceivePort| {
             Receive {
                 channel: port.channel,
             }
@@ -199,7 +206,7 @@ impl Heap {
     }
 
     fn function_run(&mut self, args: &[Pointer], responsible: Pointer) -> BuiltinResult {
-        unpack!(self, args, |closure: Closure| {
+        unpack!(self, args, |closure: &Closure| {
             closure.should_take_no_arguments()?;
             DivergeControlFlow {
                 closure: closure.address,
@@ -209,76 +216,82 @@ impl Heap {
     }
 
     fn get_argument_count(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |closure: Closure| {
-            Return(self.create_int(closure.num_args.into()))
+        unpack_and_later_drop!(self, args, |closure: &Closure| {
+            let num_args = closure.num_args.into();
+            Return(self.create_int(num_args))
         })
     }
 
     fn if_else(&mut self, args: &[Pointer], responsible: Pointer) -> BuiltinResult {
         unpack!(self, args, |condition: bool,
-                             then: Closure,
-                             else_: Closure| {
+                             then: &Closure,
+                             else_: &Closure| {
             let (run, dont_run) = if *condition {
                 (then, else_)
             } else {
                 (else_, then)
             };
-            self.drop(condition.address);
-            self.drop(dont_run.address);
+
+            let condition_address = condition.address;
+            let run_address = run.address;
+            let dont_run_address = dont_run.address;
+            self.drop(condition_address);
+            self.drop(dont_run_address);
+
             DivergeControlFlow {
-                closure: run.address,
+                closure: run_address,
                 responsible,
             }
         })
     }
 
     fn int_add(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |a: Int, b: Int| {
+        unpack_and_later_drop!(self, args, |a: &Int, b: &Int| {
             Return(self.create_int(&a.value + &b.value))
         })
     }
     fn int_bit_length(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |a: Int| {
+        unpack_and_later_drop!(self, args, |a: &Int| {
             Return(self.create_int(a.value.bits().into()))
         })
     }
     fn int_bitwise_and(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |a: Int, b: Int| {
+        unpack_and_later_drop!(self, args, |a: &Int, b: &Int| {
             Return(self.create_int(&a.value & &b.value))
         })
     }
     fn int_bitwise_or(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |a: Int, b: Int| {
+        unpack_and_later_drop!(self, args, |a: &Int, b: &Int| {
             Return(self.create_int(&a.value | &b.value))
         })
     }
     fn int_bitwise_xor(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |a: Int, b: Int| {
+        unpack_and_later_drop!(self, args, |a: &Int, b: &Int| {
             Return(self.create_int(&a.value ^ &b.value))
         })
     }
     fn int_compare_to(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |a: Int, b: Int| {
+        unpack_and_later_drop!(self, args, |a: &Int, b: &Int| {
             Return(self.create_ordering(a.value.cmp(&b.value)))
         })
     }
     fn int_divide_truncating(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |dividend: Int, divisor: Int| {
+        unpack_and_later_drop!(self, args, |dividend: &Int, divisor: &Int| {
             Return(self.create_int(&dividend.value / &divisor.value))
         })
     }
     fn int_modulo(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |dividend: Int, divisor: Int| {
+        unpack_and_later_drop!(self, args, |dividend: &Int, divisor: &Int| {
             Return(self.create_int(dividend.value.mod_floor(&divisor.value)))
         })
     }
     fn int_multiply(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |factor_a: Int, factor_b: Int| {
+        unpack_and_later_drop!(self, args, |factor_a: &Int, factor_b: &Int| {
             Return(self.create_int(&factor_a.value * &factor_b.value))
         })
     }
     fn int_parse(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text| {
             let result = match BigInt::from_str(&text.value) {
                 Ok(int) => Ok(self.create_int(int)),
                 Err(err) => Err(self.create_text(format!("{err}"))),
@@ -287,37 +300,38 @@ impl Heap {
         })
     }
     fn int_remainder(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |dividend: Int, divisor: Int| {
+        unpack_and_later_drop!(self, args, |dividend: &Int, divisor: &Int| {
             Return(self.create_int(&dividend.value % &divisor.value))
         })
     }
     fn int_shift_left(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |value: Int, amount: Int| {
+        unpack_and_later_drop!(self, args, |value: &Int, amount: &Int| {
             let amount = amount.value.to_u128().unwrap();
             Return(self.create_int(&value.value << amount))
         })
     }
     fn int_shift_right(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |value: Int, amount: Int| {
+        unpack_and_later_drop!(self, args, |value: &Int, amount: &Int| {
             let amount = amount.value.to_u128().unwrap();
             Return(self.create_int(&value.value >> amount))
         })
     }
     fn int_subtract(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |minuend: Int, subtrahend: Int| {
+        unpack_and_later_drop!(self, args, |minuend: &Int, subtrahend: &Int| {
             Return(self.create_int(&minuend.value - &subtrahend.value))
         })
     }
 
     fn list_filled(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |length: Int, item: Any| {
+        unpack_and_later_drop!(self, args, |length: &Int, item: Any| {
             let length = length.value.to_usize().unwrap();
+            let item_address = item.address;
             self.dup_by(item.address, length);
-            Return(self.create_list(vec![item.address; length]))
+            Return(self.create_list(vec![item_address; length]))
         })
     }
     fn list_get(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |list: List, index: Int| {
+        unpack_and_later_drop!(self, args, |list: &List, index: &Int| {
             let index = index.value.to_usize().unwrap();
             let item = list.items[index];
             self.dup(item);
@@ -325,23 +339,24 @@ impl Heap {
         })
     }
     fn list_insert(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |list: List, index: Int, item: Any| {
+        unpack_and_later_drop!(self, args, |list: &List, index: &Int, item: Any| {
             let mut new_list = list.items.clone();
 
             let index = index.value.to_usize().unwrap();
+            let item_address = item.address;
             self.dup(item.address);
-            new_list.insert(index, item.address);
+            new_list.insert(index, item_address);
 
             Return(self.create_list(new_list))
         })
     }
     fn list_length(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |list: List| {
+        unpack_and_later_drop!(self, args, |list: &List| {
             Return(self.create_int(list.items.len().into()))
         })
     }
     fn list_remove_at(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |list: List, index: Int| {
+        unpack_and_later_drop!(self, args, |list: &List, index: &Int| {
             let mut new_list = list.items.clone();
 
             let index = index.value.to_usize().unwrap();
@@ -351,19 +366,20 @@ impl Heap {
         })
     }
     fn list_replace(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |list: List, index: Int, new_item: Any| {
+        unpack_and_later_drop!(self, args, |list: &List, index: &Int, new_item: Any| {
             let mut new_list = list.items.clone();
 
             let index = index.value.to_usize().unwrap();
+            let new_item_address = new_item.address;
             self.dup(new_item.address);
-            new_list[index] = new_item.address;
+            new_list[index] = new_item_address;
 
             Return(self.create_list(new_list))
         })
     }
 
     fn parallel(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack!(self, args, |body_taking_nursery: Closure| {
+        unpack!(self, args, |body_taking_nursery: &Closure| {
             if body_taking_nursery.num_args != 1 {
                 return Err("`parallel` expects a closure taking a nursery.".to_string());
             }
@@ -381,7 +397,7 @@ impl Heap {
     }
 
     fn struct_get(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |struct_: Struct, key: Any| {
+        unpack_and_later_drop!(self, args, |struct_: &Struct, key: Any| {
             match struct_.get(self, key.address) {
                 Some(value) => {
                     self.dup(value);
@@ -395,21 +411,21 @@ impl Heap {
         })
     }
     fn struct_get_keys(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |struct_: Struct| {
+        unpack_and_later_drop!(self, args, |struct_: &Struct| {
             Return(self.create_list(struct_.iter().map(|(key, _)| key).collect_vec()))
         })
     }
     fn struct_has_key(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |struct_: Struct, key: Any| {
+        unpack_and_later_drop!(self, args, |struct_: &Struct, key: Any| {
             let has_key = struct_.get(self, key.address).is_some();
             Return(self.create_bool(has_key))
         })
     }
 
     fn text_characters(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text| {
+            let text = text.value.clone(); // FIXME: remove clone
             let character_addresses = text
-                .value
                 .graphemes(true)
                 .map(|it| self.create_text(it.to_string()))
                 .collect_vec();
@@ -417,22 +433,22 @@ impl Heap {
         })
     }
     fn text_concatenate(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text_a: Text, text_b: Text| {
+        unpack_and_later_drop!(self, args, |text_a: &Text, text_b: &Text| {
             Return(self.create_text(format!("{}{}", text_a.value, text_b.value)))
         })
     }
     fn text_contains(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text, pattern: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text, pattern: &Text| {
             Return(self.create_bool(text.value.contains(&pattern.value)))
         })
     }
     fn text_ends_with(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text, suffix: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text, suffix: &Text| {
             Return(self.create_bool(text.value.ends_with(&suffix.value)))
         })
     }
     fn text_from_utf8(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |bytes: List| {
+        unpack_and_later_drop!(self, args, |bytes: &List| {
             let bytes = bytes
                 .items
                 .iter()
@@ -453,7 +469,7 @@ impl Heap {
         unpack_and_later_drop!(
             self,
             args,
-            |text: Text, start_inclusive: Int, end_exclusive: Int| {
+            |text: &Text, start_inclusive: &Int, end_exclusive: &Int| {
                 let start_inclusive = start_inclusive.value.to_usize().expect(
                     "Tried to get a range from a text with an index that's too large for usize.",
                 );
@@ -471,28 +487,28 @@ impl Heap {
         )
     }
     fn text_is_empty(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text| {
             Return(self.create_bool(text.value.is_empty()))
         })
     }
     fn text_length(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text| {
             let length = text.value.graphemes(true).count().into();
             Return(self.create_int(length))
         })
     }
     fn text_starts_with(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text, prefix: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text, prefix: &Text| {
             Return(self.create_bool(text.value.starts_with(&prefix.value)))
         })
     }
     fn text_trim_end(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text| {
             Return(self.create_text(text.value.trim_end().to_string()))
         })
     }
     fn text_trim_start(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack_and_later_drop!(self, args, |text: Text| {
+        unpack_and_later_drop!(self, args, |text: &Text| {
             Return(self.create_text(text.value.trim_start().to_string()))
         })
     }
@@ -505,7 +521,7 @@ impl Heap {
     }
 
     fn try_(&mut self, args: &[Pointer]) -> BuiltinResult {
-        unpack!(self, args, |body: Closure| { Try { body: body.address } })
+        unpack!(self, args, |body: &Closure| { Try { body: body.address } })
     }
 
     fn type_of(&mut self, args: &[Pointer]) -> BuiltinResult {
@@ -548,21 +564,21 @@ impl<T> Deref for UnpackedData<T> {
     }
 }
 
-struct Any {
-    data: Data,
+struct Any<'a> {
+    data: &'a Data,
 }
-impl Deref for Any {
+impl<'a> Deref for Any<'a> {
     type Target = Data;
 
     fn deref(&self) -> &Self::Target {
-        &self.data
+        self.data
     }
 }
 
-impl TryInto<Any> for Data {
+impl<'a> TryInto<Any<'a>> for &'a Data {
     type Error = String;
 
-    fn try_into(self) -> Result<Any, Self::Error> {
+    fn try_into(self) -> Result<Any<'a>, Self::Error> {
         Ok(Any { data: self })
     }
 }