… Cargo.lock
… compiler/Cargo.toml
… compiler/src/vm/mod.rs