<!-- Please enter the corresponding issue ID: -->

<!-- Add the breaking label (PR: BREAKING) if applicable. -->

<!-- Please summarize your changes: -->
These performance improvements affect both the compilation and execution time. Detailed information can be found in the commit messages. Here's an overview of `hyperfine "cargo run --release --manifest-path=compiler/Cargo.toml -- run packages/examples/fibonacci.candy"` of Fibonacci calculation after these optimizations:

`fibonacci 8`, initial

```
Time (mean ± σ):      1.001 s ±  0.049 s    [User: 0.928 s, System: 0.066 s]
Range (min … max):    0.962 s …  1.099 s    10 runs
```

HIR/AST/CST ID mapping `Arc`

```
Time (mean ± σ):     933.9 ms ±  25.9 ms    [User: 867.5 ms, System: 65.6 ms]
Range (min … max):   904.3 ms … 973.4 ms    10 runs
```

Then, I changed the calculation to `fibonacci 15`.


```
Time (mean ± σ):      2.428 s ±  0.120 s    [User: 2.343 s, System: 0.076 s]
Range (min … max):    2.279 s …  2.658 s    10 runs
```

Channel refcounts

```
Time (mean ± σ):      1.584 s ±  0.092 s    [User: 1.501 s, System: 0.075 s]
Range (min … max):    1.499 s …  1.820 s    10 runs
```

Duplicate lookup in `dup_by`

```
Time (mean ± σ):      1.406 s ±  0.075 s    [User: 1.335 s, System: 0.070 s]
Range (min … max):    1.349 s …  1.605 s    10 runs
```

Cloning arguments in builtin functions

```
Time (mean ± σ):      1.342 s ±  0.062 s    [User: 1.268 s, System: 0.071 s]
Range (min … max):    1.273 s …  1.471 s    10 runs
```

Duplicate lookup in CSE

```
Time (mean ± σ):      1.310 s ±  0.062 s    [User: 1.238 s, System: 0.070 s]
Range (min … max):    1.235 s …  1.418 s    10 runs
```



### Checklist
<!-- Please check if your PR fulfills the following requirements: -->

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)